requirements.txt
streamlit
Pillow
requests
transformers
torch
timm
scipy
